requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
matplotlib
scipy
opencv-contrib-python
scikit-image
flask
flask-cors
pillow
boto3